fix SEC error 403: add user-agent header to EDGAR request to comply with the "Fair access" guidelines in  https://www.sec.gov/os/accessing-edgar-data:
"The SEC does not allow botnets or automated tools to crawl the site. Any request that has been identified as part of a botnet or an automated tool outside of the acceptable policy will be managed to ensure fair access for all users.

Please declare your user agent in request headers"